separate PreParseCommand into RegisterCommand
